test: converting courses to ics event format
