+++ apps/antalmanac/src/components/Calendar/Toolbar/ExportCalendar.tsx
@@ -1,271 +1,7 @@
 import { Tooltip } from '@material-ui/core';
 import Button from '@material-ui/core/Button';
 import Today from '@material-ui/icons/Today';
-import { saveAs } from 'file-saver';
-import { createEvents } from 'ics';
-
-import { HourMinute, WebsocSectionFinalExam } from 'peterportal-api-next-types';
-import { openSnackbar } from '$actions/AppStoreActions';
-import analyticsEnum, { logAnalytics } from '$lib/analytics';
-import { termData } from '$lib/termData';
-import AppStore from '$stores/AppStore';
-
-const quarterStartDates = Object.fromEntries(
-    termData
-        .filter((term) => term.startDate !== undefined)
-        .map((term) => [term.shortName, term.startDate as [number, number, number]])
-);
-const daysOfWeek = ['Su', 'M', 'Tu', 'W', 'Th', 'F', 'Sa'] as const;
-const daysOffset: Record<string, number> = { SU: -1, MO: 0, TU: 1, WE: 2, TH: 3, FR: 4, SA: 5 };
-const fallDaysOffset: Record<string, number> = { TH: 0, FR: 1, SA: 2, SU: 3, MO: 4, TU: 5, WE: 6 };
-const translateDaysForIcs = { Su: 'SU', M: 'MO', Tu: 'TU', W: 'WE', Th: 'TH', F: 'FR', Sa: 'SA' };
-const vTimeZoneSection =
-    'BEGIN:VTIMEZONE\n' +
-    'TZID:America/Los_Angeles\n' +
-    'X-LIC-LOCATION:America/Los_Angeles\n' +
-    'BEGIN:DAYLIGHT\n' +
-    'TZOFFSETFROM:-0800\n' +
-    'TZOFFSETTO:-0700\n' +
-    'TZNAME:PDT\n' +
-    'DTSTART:19700308T020000\n' +
-    'RRULE:FREQ=YEARLY;BYMONTH=3;BYDAY=2SU\n' +
-    'END:DAYLIGHT\n' +
-    'BEGIN:STANDARD\n' +
-    'TZOFFSETFROM:-0700\n' +
-    'TZOFFSETTO:-0800\n' +
-    'TZNAME:PST\n' +
-    'DTSTART:19701101T020000\n' +
-    'RRULE:FREQ=YEARLY;BYMONTH=11;BYDAY=1SU\n' +
-    'END:STANDARD\n' +
-    'END:VTIMEZONE\n' +
-    'BEGIN:VEVENT';
-
-/** [YEAR, MONTH, DAY, HOUR, MINUTE]*/
-type DateTimeArray = [number, number, number, number, number];
-/** [YEAR, MONTH, DAY]*/
-type YearMonthDay = [number, number, number];
-
-/** getByDays returns the days that a class occurs
-    Given a string of days, convert it to a list of days in ics format
-    Ex: ("TuThF") -> ["TU", "TH", "FR"] */
-const getByDays = (days: string) => {
-    return daysOfWeek.filter((day) => days.includes(day)).map((day) => translateDaysForIcs[day]);
-};
-
-/** getClassStartDate returns the start date of a class
-    Given the term and bydays, this computes the start date of the class
-    Ex: ("2021 Spring", 'Tu') -> [2021, 3, 30] */
-const getClassStartDate = (term: string, bydays: ReturnType<typeof getByDays>) => {
-    // Get the start date of the quarter (Monday)
-    const quarterStartDate = new Date(...quarterStartDates[term]);
-
-    // dayOffset represents the number of days since the start of the quarter
-    let dayOffset;
-    if (getQuarter(term) === 'Fall') {
-        // Since Fall quarter starts on a Thursday the first byday and offset
-        // will be different from other quarters
-        bydays.sort((day1, day2) => {
-            // Sorts bydays to match this ordering: [TH, FR, SA, SU, MO, TU, WE]
-            return fallDaysOffset[day1] - fallDaysOffset[day2];
-        });
-        dayOffset = fallDaysOffset[bydays[0]];
-    } else {
-        dayOffset = daysOffset[bydays[0]];
-    }
-
-    // Add the dayOffset to the quarterStartDate
-    // Date object will handle potential overflow into the next month
-    quarterStartDate.setDate(quarterStartDate.getDate() + dayOffset);
-
-    // Return [Year, Month, Date]
-    return dateToIcs(quarterStartDate);
-};
-
-/** dateToIcs takes a Date object and returns it in ics format [YYYY, MM, DD] */
-const dateToIcs = (date: Date) => {
-    return [
-        date.getFullYear(),
-        date.getMonth() + 1, // Add 1 month since it is 0-indexed
-        date.getDate(),
-    ] as YearMonthDay;
-};
-
-/** getFirstClass returns the start and end datetime of the first class
-    Ex: ([2021, 3, 30], { hour: 16, minute: 0}) -> [[2021, 3, 30, 16, 0], [2021, 3, 30, 16, 50]] */
-const getFirstClass = (
-    date: YearMonthDay,
-    startTime: HourMinute,
-    endTime: HourMinute
-): [DateTimeArray, DateTimeArray] => {
-    const [classStartTime, classEndTime] = parseTimes(startTime, endTime);
-    return [
-        [...date, ...classStartTime],
-        [...date, ...classEndTime],
-    ];
-};
-
-/** getExamTime returns the start and end datetime of an exam
-    Ex: (WebsocSectionFinalExam, "2019") -> [[2019, 6, 7, 10, 30], [2019, 6, 7, 12, 30]] */
-const getExamTime = (exam: WebsocSectionFinalExam, year: number) => {
-    if (exam.month && exam.day && exam.startTime && exam.endTime) {
-        const month = exam.month;
-        const day = exam.day;
-        const [examStartTime, examEndTime] = parseTimes(exam.startTime, exam.endTime);
-
-        return [
-            [year, month + 1, day, ...examStartTime],
-            [year, month + 1, day, ...examEndTime],
-        ];
-    } else {
-        // This should never happen, but we return an empty array for typescript purposes
-        return [];
-    }
-};
-
-/** parseTimes converts startTime and endTime
-    This is a helper function used by getFirstClass and getExamTime
-    Ex: { hour: 16, minute: 0}, { hour: 16, minute: 50 } -> [[16, 0], [16, 50]] */
-const parseTimes = (startTime: HourMinute, endTime: HourMinute) => {
-    return [
-        [startTime.hour, startTime.minute],
-        [endTime.hour, endTime.minute],
-    ] as const;
-};
-
-/** getYear returns the year of a given term
-    Ex: "2019 Fall" -> "2019" */
-const getYear = (term: string) => {
-    return parseInt(term.split(' ')[0]);
-};
-
-/** getQuarter returns the quarter of a given term
-    Ex: "2019 Fall" -> "Fall" */
-const getQuarter = (term: string) => {
-    return term.split(' ')[1];
-};
-
-// getTermLength returns the number of weeks in a given term,
-// which is 10 for quarters and Summer Session 10wk,
-// and 5 for Summer Sessions I and II
-const getTermLength = (quarter: string) => (quarter.startsWith('Summer') && quarter !== 'Summer10wk' ? 5 : 10);
-
-// getRRule returns a string representing the recurring rule for the VEvent
-//  Ex: ["TU", "TH"] -> "FREQ=WEEKLY;BYDAY=TU,TH;INTERVAL=1;COUNT=20"
-const getRRule = (bydays: ReturnType<typeof getByDays>, quarter: string) => {
-    let count = getTermLength(quarter) * bydays.length; // Number of occurences in the quarter
-    switch (quarter) {
-        case 'Fall':
-            for (const byday of bydays) {
-                switch (byday) {
-                    case 'TH':
-                    case 'FR':
-                    case 'SA':
-                        count += 1; // account for Week 0 course meetings
-                        break;
-                    default:
-                        break;
-                }
-            }
-            break;
-        case 'Summer1':
-            if (bydays.includes('MO')) count += 1; // instruction ends Monday of Week 6
-            break;
-        case 'Summer10wk':
-            if (bydays.includes('FR')) count -= 1; // instruction ends Thursday of Week 10
-            break;
-        default:
-            break;
-    }
-    return `FREQ=WEEKLY;BYDAY=${bydays.toString()};INTERVAL=1;COUNT=${count}`;
-};
-
-const exportCalendar = () => {
-    // Fetch courses for the current schedule
-    const courses = AppStore.schedule.getCurrentCourses();
-
-    // Construct an array of VEvents for each event
-    const events = [];
-    for (const course of courses) {
-        const {
-            term,
-            deptCode,
-            courseNumber,
-            courseTitle,
-            section: { sectionType, instructors, meetings, finalExam },
-        } = course;
-
-        // Create a VEvent for each meeting
-        for (const meeting of meetings) {
-            if (meeting.timeIsTBA) {
-                // Skip this meeting if there is no meeting time
-                continue;
-            }
-
-            if (meeting.days && meeting.startTime && meeting.endTime) {
-                const bydays = getByDays(meeting.days);
-                const classStartDate = getClassStartDate(term, bydays);
-                const [firstClassStart, firstClassEnd] = getFirstClass(
-                    classStartDate,
-                    meeting.startTime,
-                    meeting.endTime
-                );
-                const rrule = getRRule(bydays, getQuarter(term));
-
-                // Add VEvent to events array
-                events.push({
-                    productId: 'antalmanac/ics',
-                    startOutputType: 'local' as const,
-                    endOutputType: 'local' as const,
-                    title: `${deptCode} ${courseNumber} ${sectionType}`,
-                    description: `${courseTitle}\nTaught by ${instructors.join('/')}`,
-                    location: `${meeting.bldg}`,
-                    start: firstClassStart as DateTimeArray,
-                    end: firstClassEnd as DateTimeArray,
-                    recurrenceRule: rrule,
-                });
-            }
-        }
-
-        // Add Final to events
-        if (finalExam && finalExam.examStatus == 'SCHEDULED_FINAL') {
-            const [examStart, examEnd] = getExamTime(finalExam, getYear(term));
-            events.push({
-                productId: 'antalmanac/ics',
-                startOutputType: 'local' as const,
-                endOutputType: 'local' as const,
-                title: `${deptCode} ${courseNumber} Final Exam`,
-                description: `Final Exam for ${courseTitle}`,
-                start: examStart as DateTimeArray,
-                end: examEnd as DateTimeArray,
-            });
-        }
-    }
-
-    // Convert the events into a vcalendar
-    // Callback function triggers a download of the .ics file
-    createEvents(events, (err, val) => {
-        logAnalytics({
-            category: 'Calendar Pane',
-            action: analyticsEnum.calendar.actions.DOWNLOAD,
-        });
-        if (!err) {
-            // Add timezone information to start and end times for events
-            const icsString = val
-                .replaceAll('DTSTART', 'DTSTART;TZID=America/Los_Angeles')
-                .replaceAll('DTEND', 'DTEND;TZID=America/Los_Angeles');
-            // Download the .ics file
-            saveAs(
-                // inject the VTIMEZONE section into the .ics file
-                new Blob([icsString.replace('BEGIN:VEVENT', vTimeZoneSection)], { type: 'text/plain;charset=utf-8' }),
-                'schedule.ics'
-            );
-            openSnackbar('success', 'Schedule downloaded!', 5);
-        } else {
-            openSnackbar('error', 'Something went wrong! Unable to download schedule.', 5);
-            console.log(err);
-        }
-    });
-};
+import { exportCalendar } from '$lib/download';
 
 const ExportCalendarButton = () => (
     <Tooltip title="Download Calendar as an .ics file">
